backend/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2
djangorestframework-simplejwt